+++ npm-scripts/compile-js.js
@@ -1,5 +1,5 @@
 import fs from 'fs';
-import glob from 'glob';
+import { glob } from 'glob';
 import terser from '@rollup/plugin-terser';
 import dotenv from 'dotenv';
 dotenv.config({ path: `.env.${process.env.NODE_ENV === 'production' ? 'production' : 'development'}` });
@@ -7,6 +7,7 @@ dotenv.config({ path: `.env.${process.env.NODE_ENV === 'production' ? 'productio
 // envから値を取得
 const isMinify = JSON.parse(process.env.MINIFY);
 const dist = process.env.DIST;
+const argTargetFile = process.env.TARGET_FILE;
 
 const setting = (name) => {
   return {
@@ -20,10 +21,11 @@ const setting = (name) => {
 };
 
 // 出力先のJSファイルを削除
-if (process.env.NODE_ENV !== 'production') {
+if (process.env.NODE_ENV !== 'production' && !argTargetFile) {
   glob.sync(`${dist}/scripts/**/*.js`).forEach((file) => {
     fs.unlinkSync(file);
   });
 }
-const settings = glob.sync(`src/scripts/**/!(_)*.js`).map((file) => setting(file));
+const files = argTargetFile ? new Array(argTargetFile) : glob.sync(`src/scripts/**/!(_)*.js`);
+const settings = files.map((file) => setting(file));
 export default settings;
